Frontend recieving fetched data
Firstly, I've added a Components folder and created a WeatherCard component. Then, I removed the React boilerplate
code and started working on rendering the weather cards. I then implemented the fetching script to fetch data from the backend and display it on the weather cards. Basic styling is added to make the cards stand clearer.